refactor(cli): extract feedback subcommand into cli/commands/feedback.mjs
Move the `feedback` (feedback add) block out of main() into a dedicated
handler. Behavior, messages, and exit codes are unchanged; only the enclosing
function and relative import depth differ.
